Clean up Java 12 and 14 testing
