+++ org.eclipse.jdt.ui.tests/ui/org/eclipse/jdt/ui/tests/quickfix/AssistQuickFixTest14.java
@@ -58,6 +58,66 @@ public class AssistQuickFixTest14 extends QuickFixTest {
 	}
 
 	@Test
+	public void testSplitSwitchCaseLabelRuleStatement() throws Exception {
+		fJProject1= JavaProjectHelper.createJavaProject("TestProject1", "bin");
+		fJProject1.setRawClasspath(projectSetup.getDefaultClasspath(), null);
+		JavaProjectHelper.set14CompilerOptions(fJProject1, true);
+		fSourceFolder= JavaProjectHelper.addSourceContainer(fJProject1, "src");
+
+		String str= """
+			module test {
+			}
+			""";
+		IPackageFragment def= fSourceFolder.createPackageFragment("", false, null);
+		def.createCompilationUnit("module-info.java", str, false, null);
+
+		IPackageFragment pack= fSourceFolder.createPackageFragment("test", false, null);
+		String str1= """
+			package test;
+			public class Cls {
+			    public static void foo(Day day) {
+			        String weekDayOrEnd = switch (day) {
+			            case SATURDAY, SUNDAY -> "Weekend";
+			            case MONDAY, TUESDAY, WEDNESDAY -> "Weekday";
+			            case THURSDAY, FRIDAY -> "Weekday";
+			        };
+			    }
+			}
+
+			enum Day {
+			    MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY;
+			}
+			""";
+		ICompilationUnit cu= pack.createCompilationUnit("Cls.java", str1, false, null);
+
+		IInvocationContext ctx= getCorrectionContext(cu, 239, 0);
+		ArrayList<IJavaCompletionProposal> proposals= collectAssists(ctx, false);
+
+		CUCorrectionProposal proposal= (CUCorrectionProposal) proposals.get(0);
+		String preview= getPreviewContent(proposal);
+
+		String expected= """
+			package test;
+			public class Cls {
+			    public static void foo(Day day) {
+			        String weekDayOrEnd = switch (day) {
+			            case SATURDAY, SUNDAY -> "Weekend";
+			            case MONDAY, TUESDAY, WEDNESDAY -> "Weekday";
+			            case THURSDAY -> "Weekday";
+						case FRIDAY -> "Weekday";
+			        };
+			    }
+			}
+
+			enum Day {
+			    MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY;
+			}
+			""";
+
+		assertEqualStringsIgnoreOrder(new String[] { preview }, new String[] { expected });
+	}
+
+	@Test
 	public void testConvertToSwitchExpression1() throws Exception {
 		fJProject1= JavaProjectHelper.createJavaProject("TestProject1", "bin");
 		fJProject1.setRawClasspath(projectSetup.getDefaultClasspath(), null);
